requirements.txt
authorizenet==1.1.0
certifi==2018.8.24
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.5
django-cors-headers==2.2.0
django-filter==1.1.0
django-heroku==0.3.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.7
lxml==4.2.5
Markdown==2.6.11
pkg-resources==0.0.0
psycopg2==2.7.5
PyJWT==1.6.3
pytz==2018.4
PyXB==1.2.5
requests==2.21.0
stripe==2.20.3
urllib3==1.23
whitenoise==4.1
